refactor(evo_git): extract Slots and Worktrees helper modules from AgentScheduler
Break down the monolithic agent_scheduler.ex (~1420 lines) by extracting
two logical groups into dedicated helper modules:

1. EvoGit.AgentScheduler.Slots (slots.ex) — LLM and tool slot management
   including backoff logic, pending queue grants, and retry timers.
   GenServer callbacks delegate to pure functions in this module.

2. EvoGit.AgentScheduler.Worktrees (worktrees.ex) — Worktree lifecycle
   management including initialization, assignment, preparation, init
   script execution, deletion, orphaned branch cleanup, and commit sync.

The AgentScheduler module retains all public API signatures, GenServer
callbacks, agent lifecycle, subagent spawning, and ETS helpers. No
behavioral changes — all 192 tests pass.
